auth and check user_id

+++ src/app/Http/Controllers/Tweet/DeleteController.php
@@ -4,16 +4,24 @@ namespace App\Http\Controllers\Tweet;
 
 use App\Http\Controllers\Controller;
 use App\Models\Tweet;
+use App\Services\TweetService;
 use Illuminate\Http\Request;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class DeleteController extends Controller
 {
+    public function __construct(private readonly TweetService $service)
+    {
+    }
     /**
      * Handle the incoming request.
      */
     public function __invoke(Request $request)
     {
         $tweetId = (int) $request->route('tweetId');
+        if ($this->service->checkOwnTweet($request->user()->id, $tweetId)) {
+            throw new AccessDeniedHttpException();
+        }
         $tweet = Tweet::where('id', $tweetId)->firstOrFail();
         $tweet->delete();
 //        Tweet::destroy($tweetId);
